6. Идеальный вес [#41584]

+++ src/ru/job4j/converter/Converter.java
@@ -14,7 +14,7 @@ public class Converter {
 
     public static void main(String[] args) {
         int euro = Converter.rubleToEuro(140);
-        float dollar = Converter.rubleToDollar(140);
+        int dollar = Converter.rubleToDollar(140);
         System.out.println("140 рублей это " + euro + " euro при курсе 70р за 1 евро");
         System.out.println("140 рублей это " + dollar + " dollar при курсе 60р за один евро, но это не точно :)");
     }
